+++ pkg/controller/util/genericinformer.go
@@ -21,11 +21,11 @@ import (
 
 	"github.com/pkg/errors"
 
+	"github.com/kubernetes-sigs/federation-v2/pkg/client/generic/scheme"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	pkgruntime "k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/watch"
-	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/cache"
 	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"