+++ scripts/pre-commit.sh
@@ -48,16 +48,12 @@ function download-dependencies() {
 }
 
 function run-unit-tests() {
-  # Ensure the test binaries are in the path.
-  export TEST_ASSET_PATH="${base_dir}/bin"
-  export TEST_ASSET_ETCD="${TEST_ASSET_PATH}/etcd"
-  export TEST_ASSET_KUBE_APISERVER="${TEST_ASSET_PATH}/kube-apiserver"
   ${MAKE_CMD} test
 }
 
 function run-e2e-tests-with-managed-fixture() {
   # Ensure the test binaries are in the path.
-  export TEST_ASSET_PATH="${base_dir}/bin"
+  export TEST_ASSET_PATH="${ROOT_DIR}/bin"
   export TEST_ASSET_ETCD="${TEST_ASSET_PATH}/etcd"
   export TEST_ASSET_KUBE_APISERVER="${TEST_ASSET_PATH}/kube-apiserver"
   ${MANAGED_E2E_TEST_CMD}
@@ -85,18 +81,9 @@ function run-namespaced-e2e-tests-with-unmanaged-fixture() {
   ${namespaced_e2e_test_cmd} --ginkgo.focus=Placement
 }
 
-function check-kubebuilder-output() {
-  ./bin/kubebuilder generate
-  echo "Checking state of working tree after running 'kubebuilder generate'"
-  check-git-state
-}
-
-function check-install-yaml() {
-  PATH="${PATH}:${base_dir}/bin" FEDERATION_NAMESPACE=federation-system \
-    INSTALL_YAML=./hack/install-latest.yaml \
-    ./scripts/generate-install-yaml.sh \
-    ${CONTAINER_REGISTRY_HOST}/federation-v2:latest
-  echo "Checking state of working tree after generating install yaml"
+function check-make-generate-output() {
+  ${MAKE_CMD} generate
+  echo "Checking state of working tree after running 'make generate'"
   check-git-state
 }
 
@@ -115,11 +102,12 @@ function check-git-state() {
 
 # Make sure, we run in the root of the repo and
 # therefore run the tests on all packages
-base_dir="$( cd "$(dirname "$0")/.." && pwd )"
-cd "$base_dir" || {
-  echo "Cannot cd to '$base_dir'. Aborting." >&2
+cd "$ROOT_DIR" || {
+  echo "Cannot cd to '$ROOT_DIR'. Aborting." >&2
   exit 1
 }
+
+export PATH=${ROOT_DIR}/bin:${PATH}
 
 echo "Downloading test dependencies"
 download-dependencies
@@ -133,23 +121,14 @@ echo "Verifying Gofmt"
 echo "Checking that correct Error Package is used."
 ./hack/verify-errpkg.sh
 
-echo "Checking that 'kubebuilder generate' is up-to-date"
-check-kubebuilder-output
-
-echo "Checking that hack/install-latest.yaml is up-to-date"
-check-install-yaml
+echo "Checking that 'make generate' is up-to-date"
+check-make-generate-output
 
 echo "Checking that fixture is available for all federate directives"
 ./scripts/check-directive-fixtures.sh
 
 echo "Building federation binaries"
 build-binaries
-
-echo "Checking sync up status of helm chart"
-BUILD_KUBEFED="false" PATH="${PATH}:${base_dir}/bin" ./scripts/sync-up-helm-chart.sh
-
-echo "Checking helm chart state of working tree"
-check-git-state
 
 echo "Running unit tests"
 run-unit-tests
@@ -160,43 +139,26 @@ run-e2e-tests-with-managed-fixture
 echo "Downloading e2e test dependencies"
 ./scripts/download-e2e-binaries.sh
 
-export PATH=${TEST_ASSET_PATH}:${PATH}
-
 CREATE_INSECURE_REGISTRY=y CONFIGURE_INSECURE_REGISTRY=y OVERWRITE_KUBECONFIG=y \
     ./scripts/create-clusters.sh
 
 # Initialize list of clusters to join
 join-cluster-list > /dev/null
 
-echo "Deploying cluster-scoped federation-v2 with script"
+echo "Deploying cluster-scoped federation-v2"
 ./scripts/deploy-federation.sh ${CONTAINER_REGISTRY_HOST}/federation-v2:e2e $(join-cluster-list)
 
-echo "Running go e2e tests with cluster-scoped unmanaged fixture deployed with script"
+echo "Running go e2e tests with cluster-scoped unmanaged fixture"
 run-e2e-tests-with-unmanaged-fixture
 
 echo "Deleting cluster-scoped federation-v2"
 ./scripts/delete-federation.sh
 
-echo "Deploying cluster-scoped federation-v2 with helm chart"
-USE_CHART=y ./scripts/deploy-federation.sh ${CONTAINER_REGISTRY_HOST}/federation-v2:e2e $(join-cluster-list)
-
-echo "Running go e2e tests with cluster-scoped unmanaged fixture deployed by helm chart"
-run-e2e-tests-with-unmanaged-fixture
-
-echo "Deleting cluster-scoped federation-v2 with helm chart"
-USE_CHART=y ./scripts/delete-federation.sh
-
 echo "Deploying namespace-scoped federation-v2"
 FEDERATION_NAMESPACE=foo NAMESPACED=y ./scripts/deploy-federation.sh ${CONTAINER_REGISTRY_HOST}/federation-v2:e2e $(join-cluster-list)
 
-echo "Running go e2e tests with namespace-scoped unmanaged fixture deployed by script"
+echo "Running go e2e tests with namespace-scoped unmanaged fixture"
 run-namespaced-e2e-tests-with-unmanaged-fixture
 
 echo "Deleting namespace-scoped federation-v2"
 FEDERATION_NAMESPACE=foo NAMESPACED=y DELETE_CLUSTER_RESOURCE=y ./scripts/delete-federation.sh
-
-echo "Deploying namespaced-scoped federation-v2 with helm chart"
-FEDERATION_NAMESPACE=foo NAMESPACED=y USE_CHART=y ./scripts/deploy-federation.sh ${CONTAINER_REGISTRY_HOST}/federation-v2:e2e $(join-cluster-list)
-
-echo "Running go e2e tests with namespace-scoped unmanaged fixture deployed by helm chart"
-run-namespaced-e2e-tests-with-unmanaged-fixture